modules/pam_tally, modules/pam_tally2: fix compilation warnings
Fix the following compilation warnings reported by gcc
when sizeof(time_t) > sizeof(long), e.g. on x32:

  modules/pam_tally/pam_tally.c:541:7: warning: format ‘%ld’ expects argument of type ‘long int’, but argument 5 has type ‘time_t’ {aka ‘long long int’} [-Wformat=]
    541 |     _("The account is temporarily locked (%ld seconds left)."),
        |       ^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
  modules/pam_tally/pam_tally.c:546:40: warning: format ‘%ld’ expects argument of type ‘long int’, but argument 6 has type ‘time_t’ {aka ‘long long int’} [-Wformat=]
    546 |       "user %s (%lu) has time limit [%lds left]"
        |                                      ~~^
        |                                        |
        |                                        long int
        |                                      %lld
  ......
    549 |       oldtime+lock_time-time(NULL));
        |       ~~~~~~~~~~~~~~~~~~~~~~~~~~~~
        |                        |
        |                        time_t {aka long long int}

  modules/pam_tally2/pam_tally2.c:592:27: warning: format ‘%ld’ expects argument of type ‘long int’, but argument 5 has type ‘time_t’ {aka ‘long long int’} [-Wformat=]
    592 |          pam_info(pamh, _("The account is temporarily locked (%ld seconds left)."),
        |                           ^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
  modules/pam_tally2/pam_tally2.c:597:50: warning: format ‘%ld’ expects argument of type ‘long int’, but argument 6 has type ‘time_t’ {aka ‘long long int’} [-Wformat=]
    597 |                 "user %s (%lu) has time limit [%lds left]"
        |                                                ~~^
        |                                                  |
        |                                                  long int
        |                                                %lld
  ......
    600 |                 oldtime+opts->lock_time-time(NULL));
        |                 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
        |                                        |
        |                                        time_t {aka long long int}

This change doesn't attempt to fix handling of 64-bit time_t on 32-bit
systems in these modules.

* modules/pam_tally/pam_tally.c (tally_check): Cast time_t expressions
to long int before passing them to pam_info and pam_syslog.
* modules/pam_tally2/pam_tally2.c (tally_check): Likewise.

+++ modules/pam_tally/pam_tally.c
@@ -539,14 +539,14 @@ tally_check (time_t oldtime, pam_handle_t *pamh, uid_t uid,
 	  if (!(opts->ctrl & OPT_SILENT))
 	       pam_info (pamh,
 			 _("The account is temporarily locked (%ld seconds left)."),
-			 oldtime+lock_time-time(NULL));
+			 (long int) (oldtime+lock_time-time(NULL)));
 
 	  if (!(opts->ctrl & OPT_NOLOGNOTICE))
 	       pam_syslog (pamh, LOG_NOTICE,
 			   "user %s (%lu) has time limit [%lds left]"
 			   " since last failure.",
 			   user, (unsigned long int) uid,
-			   oldtime+lock_time-time(NULL));
+			   (long int) (oldtime+lock_time-time(NULL)));
 		return PAM_AUTH_ERR;
 	}
       }
